requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
black==25.9.0
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
curlify==3.0.0
Django==5.2.7
django-cors-headers==4.9.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
dnspython==2.8.0
et_xmlfile==2.0.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
lxml==6.0.2
mypy_extensions==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prettytable==3.16.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pymongo==4.15.3
PyMuPDF==1.26.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytokens==0.2.0
pytz==2025.2
reportlab==4.4.4
requests==2.32.5
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
toml==0.10.2
tomli==2.3.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
xlsxwriter==3.2.9
